refactor(orb-ui): Refactor Orb Entities
Create Orb Entity Base interface for common properties shared among all Orb entities.
Current properties being name, id and mf_owner_id, as described in the [wiki](https://github.com/ns1labs/orb/wiki/Architecture:-Common-Patterns)

Refactor interfaces to allow for easier maintenance
