+++ ui/src/app/common/interfaces/orb/agent.interface.ts
@@ -4,20 +4,12 @@
  * [Fleet Architecture]{@link https://github.com/ns1labs/orb/wiki/Architecture:-Fleet}
  */
 
+import { OrbEntity } from 'app/common/interfaces/orb/orb.entity.interface';
+
 /**
  * @interface Agent
  */
-export interface Agent {
-  /**
-   *  ID {string} UUIDv4 (read only)
-   */
-  id?: string;
-
-  /**
-   * Name {string} [a-zA-Z_:][a-zA-Z0-9_]*
-   */
-  name?: string;
-
+export interface Agent extends OrbEntity {
   /**
    * A timestamp of creation {string}
    */
@@ -64,4 +56,10 @@ export interface Agent {
    * Last Heartbeat timestamp {string}
    */
   ts_lst_hb?: string;
+
+  /**
+   * Boolean which indicates whether the agent is in an error state or not.
+   * Heartbeat data contains error information.
+   */
+  error_state?: boolean;
 }
