+++ ui/src/app/common/interfaces/orb/dataset.interface.ts
@@ -4,25 +4,12 @@
  * [Dataset Architecture]{@link https://github.com/ns1labs/orb/wiki/Architecture:-Policies-and-Datasets}
  */
 
+import { OrbEntity } from 'app/common/interfaces/orb/orb.entity.interface';
+
 /**
  * @interface Dataset
  */
-export interface Dataset {
-  /**
-   *  Tenant owner {string} UUIDv4 (read only)
-   */
-  mf_owner_id?: string;
-
-  /**
-   *  ID {string} UUIDv4 (read only)
-   */
-  id?: string;
-
-  /**
-   * Name {string} [a-zA-Z_:][a-zA-Z0-9_]*
-   */
-  name?: string;
-
+export interface Dataset extends OrbEntity {
   /**
    *  Agent Group ID {string} UUIDv4 (read only)
    */
@@ -58,4 +45,5 @@ export interface Dataset {
    * Dataset Metadata {{[propName: string]: string}}
    */
   dataset_metadata?: any;
+
 }
