+++ ui/src/app/common/interfaces/orb/orb.entity.interface.ts
@@ -1,0 +1,26 @@
+/**
+ * Base Orb Entity Data Model Interface
+ *
+ * [Orb Base Entity]{@link https://github.com/ns1labs/orb/wiki/Architecture:-Common-Patterns}
+ */
+
+/**
+ * @interface OrbEntity
+ */
+export interface OrbEntity {
+  /**
+   *  ID {string} UUIDv4 (read only)
+   */
+  id?: string;
+
+  /**
+   * Name {string} [a-zA-Z_:][a-zA-Z0-9_]*
+   */
+  name?: string;
+
+  /**
+   *  Tenant owner {string} UUIDv4 (read only)
+   */
+  mf_owner_id?: string;
+
+}
