+++ ui/src/app/pages/fleet/agents/list/agent.list.component.ts
@@ -5,7 +5,7 @@ import { DropdownFilterItem } from 'app/common/interfaces/mainflux.interface';
 import { ActivatedRoute, Router } from '@angular/router';
 import { STRINGS } from 'assets/text/strings';
 import { ColumnMode, TableColumn } from '@swimlane/ngx-datatable';
-import { NgxDatabalePageInfo, OrbPagination } from 'app/common/interfaces/orb/pagination';
+import { NgxDatabalePageInfo, OrbPagination } from 'app/common/interfaces/orb/pagination.interface';
 import { Debounce } from 'app/shared/decorators/utils';
 import { Agent } from 'app/common/interfaces/orb/agent.interface';
 import { AgentsService } from 'app/common/services/agents/agents.service';
@@ -190,5 +190,5 @@ export class AgentListComponent implements OnInit, AfterViewInit {
     });
   }
 
-  filterByActive = (agent) => agent.state === 'active';
+  filterByError = (agent) => agent.error_state;
 }
