+++ ui/src/app/pages/sinks/list/sink.list.component.ts
@@ -7,10 +7,11 @@ import { SinkDetailsComponent } from 'app/pages/sinks/details/sink.details.compo
 import { ActivatedRoute, Router } from '@angular/router';
 import { STRINGS } from 'assets/text/strings';
 import { ColumnMode, TableColumn } from '@swimlane/ngx-datatable';
-import { NgxDatabalePageInfo, OrbPagination } from 'app/common/interfaces/orb/pagination';
+import { NgxDatabalePageInfo, OrbPagination } from 'app/common/interfaces/orb/pagination.interface';
 import { AgentGroup } from 'app/common/interfaces/orb/agent.group.interface';
 import { Debounce } from 'app/shared/decorators/utils';
 import { SinkDeleteComponent } from 'app/pages/sinks/delete/sink.delete.component';
+import { Sink } from 'app/common/interfaces/orb/sink.interface';
 
 @Component({
   selector: 'ngx-sink-list-component',
